requirements.txt
websocket-client==1.3.1
pyre-gevent==0.2.3